backend/requirements.txt
openai
python-dotenv
fastapi[standard]
pymongo
openpyxl
bcrypt
pyjwt
pillow